chore: rename error.rs to errors.rs
For consistency
