+++ programs/lazorkit/src/lib.rs
@@ -1,7 +1,7 @@
 use anchor_lang::prelude::*;
 
 pub mod constants;
-pub mod error;
+pub mod errors;
 pub mod instructions;
 pub mod state;
 pub mod utils;
